method to access all conf info
